requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.3.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-yasg==1.21.7
django-cors-headers==4.3.1
requests==2.32.3
redis==5.0.6
celery==5.4.0
eventlet==0.36.1
django-celery-beat==2.6.0
loguru==0.7.2
coverage==7.5.3
pytz~=2024.1